Update repository interfaces and services to use Long instead of Integer for entity IDs

+++ src/main/java/spring/workshop/expenses/entities/Category.java
@@ -16,7 +16,7 @@ public class Category {
     @Id
     @Column(name = "category_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column
     @NonNull
@@ -30,16 +30,16 @@ public class Category {
         this.name = name;
     }
 
-    public Category(Integer id, @NonNull String name) {
+    public Category(Long id, @NonNull String name) {
         this.name = name;
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
